pyproject.toml
[project]
name = "resume-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black==24.4.2",
    "clean-text==0.6.0",
    "isort==5.13.2",
    "multiprocess==0.70.15",
    "opencv-python==4.9.0.80",
    "pdf2image==1.17.0",
    "pdfplumber==0.10.3",
    "pillow==10.0.0",
    "pytesseract==0.3.10",
    "python-dotenv==1.0.1",
    "requests==2.31.0",
    "spacy==3.7.2",
    "streamlit==1.34.0",
    "typer<0.10.0",
]
